chore: migrate ECN multidiut
