+++ tests/common/plugins/conditional_mark/tests_mark_conditions.yaml
@@ -2074,6 +2074,18 @@ snappi_tests:
     conditions:
       - "asic_type in ['vs']"
 
+snappi_tests/ecn/test_ecn_marking_with_pfc_quanta_variance_with_snappi.py:
+  skip:
+    reason: "Current test case only work with cisco platform"
+    conditions:
+      - "asic_type not in ['cisco-8000']"
+
+snappi_tests/ecn/test_ecn_marking_with_snappi.py:
+  skip:
+    reason: "Current test case only work with cisco platform"
+    conditions:
+      - "asic_type not in ['cisco-8000']"
+
 snappi_tests/ecn/test_red_accuracy_with_snappi:
   skip:
     reason: "Test should not be run as part of nightly. / Snappi test only support on physical tgen testbed"
@@ -2081,6 +2093,7 @@ snappi_tests/ecn/test_red_accuracy_with_snappi:
     conditions:
       - "topo_type in ['tgen']"
       - "asic_type in ['vs']"
+
 
 snappi_tests/multidut/pfc/test_multidut_global_pause_with_snappi.py:
   skip:
